@@ -1,11 +1,9 @@
 module.exports = (sequelize, Sequelize) => {
-	const thiscanbeanything = sequelize.define('asset', {
-	  symbol: {
-				type: Sequelize.STRING(6),
-				unique: true,
-				primaryKey: true
-	  },
-	  shares: { // total shares bought 
+	const asset = sequelize.define('asset', {
+		symbol: {
+			type: Sequelize.STRING(6)
+ 		 },
+	  	shares: { // total shares bought 
 				type: Sequelize.INTEGER,
 				defaultValue: 0,
 				validate: {
@@ -100,8 +98,15 @@ module.exports = (sequelize, Sequelize) => {
 			validate: {
 				isDecimal: true,
 			}
-		}
+		},
+		assettype : {
+			type: Sequelize.STRING(10),
+			defaultValue: 'existing',
+		},
+		ownerid: {
+			type: Sequelize.INTEGER
+ 		 },
 	});
 	
-	return thiscanbeanything;
+	return asset;
 }
