@@ -8,8 +8,6 @@ app.use(cors())
  
 const db = require('./app/config/db.config.js');
   
-
-
 // create table if it DOES NOT exist
 db.sequelize.sync().then(() => {
   console.log('Sync with table');
@@ -17,7 +15,9 @@ db.sequelize.sync().then(() => {
  
 require('./app/route/transactions.route.js')(app);
 require('./app/route/asset.route.js')(app);
- 
+require('./app/route/owners-contributions.route.js')(app);
+require('./app/route/misc.route.js')(app);
+
 // Create a Server
 var server = app.listen(8080, function () {
  
