@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import {MatDialog} from '@angular/material';
 
 @Component({
   selector: 'app-root',
@@ -8,5 +9,24 @@ import { Component } from '@angular/core';
 export class AppComponent {
   title = 'app';
 
-constructor(){}
+constructor(public dialog: MatDialog){}
+
+openContributionForm() {
+  const dialogRef = this.dialog.open(ContributionForm);
+
+  dialogRef.afterClosed().subscribe(result => {
+    console.log(`Dialog results after open: ${result}`);
+  });
 }
+}
+
+
+@Component({
+  selector: 'contribution-form',
+  templateUrl: './Forms/contribution-form.html',
+})
+export class ContributionForm {
+
+
+}
+
