@@ -25,8 +25,10 @@ export class AssetService {
   }
 
   // return all assets from the database table assets
-  getAllAssets (): Observable<asset[]> {
-    return this.http.get<asset[]>(this.Url+'currentassets')
+  getAllAssetsByType (type: string): Observable<asset[]> {
+    // type referse to the assettype: existing, archived, pure
+    const route = type == "all" ? "currentassets" : "allassets/"+type;
+    return this.http.get<asset[]>(this.Url+route)
   }
 
   // create an asset
@@ -38,4 +40,24 @@ export class AssetService {
   updateAsset (asset: asset): Observable<any> {
     return this.http.put(this.Url+'currentassets', asset, httpOptions);
   }
+
+  // delete an asset
+  deleteAsset (symbol: string) {
+    console.log(this.Url+'currentassets/'+symbol)
+    return this.http.delete(this.Url+'currentassets/'+symbol, httpOptions);
+  }
+
+  /*******************ARCHIVED ASSETS  *****************************/
+  // Create an archived Asset
+  transferToArchive(id: any, status: any) {
+    return this.http.put(this.Url+'transferAsset/'+id+"/"+status, httpOptions);
+  }
+
+  // Return a single asset from the database table assets
+  getAssetIfNotExisting(symbol: string): Observable<asset> {
+    const url = `${this.Url + 'currentExistingAssets'}/${symbol}`;
+    let existingAsset = new asset();
+    return this.http.get<asset>(url);
+  }
+  
 }
